refactor: ダウンロードデータの取得を関数に分離
